TST: Move pyca-specific test infrastructure to test_pyca
