+++ test/conftest.py
@@ -1,6 +1,4 @@
-import threading
 import logging
-import pyca
 from pcaspy import Driver, SimpleServer
 from pcaspy.tools import ServerThread
 import pytest
@@ -86,77 +84,3 @@ def server():
         server.start_server()
         yield server
         server.kill_server()
-
-
-class ConnectCallback(object):
-    def __init__(self, name):
-        self.name = name
-        self.connected = False
-        self.cev = threading.Event()
-        self.dcev = threading.Event()
-        self.lock = threading.RLock()
-
-    def wait(self, timeout=None):
-        logger.debug('Wait on connect callback %s', self.name)
-        ok = self.cev.wait(timeout=timeout)
-        if ok:
-            logger.debug('Wait complete on connect %s', self.name)
-        else:
-            logger.debug('Wait fail on connect %s', self.name)
-        return ok
-
-    def wait_dc(self, timeout=None):
-        logger.debug('Wait on disconnect callback %s', self.name)
-        ok = self.dcev.wait(timeout=timeout)
-        if ok:
-            logger.debug('Wait complete on disconnect %s', self.name)
-        else:
-            logger.debug('Wait fail on disconnect %s', self.name)
-        return ok
-
-    def __call__(self, is_connected):
-        logger.debug('Connect callback in %s, is_connected=%s',
-                     self.name, is_connected)
-        with self.lock:
-            self.connected = is_connected
-            if self.connected:
-                self.cev.set()
-                self.dcev.clear()
-            else:
-                self.dcev.set()
-                self.cev.clear()
-
-
-class GetCallback(object):
-    def __init__(self, name):
-        self.name = name
-        self.gev = threading.Event()
-
-    def wait(self, timeout=None):
-        logger.debug('Wait on get callback %s', self.name)
-        ok = self.gev.wait(timeout=timeout)
-        if ok:
-            logger.debug('Wait complete on get %s', self.name)
-        else:
-            logger.debug('Wait fail on get %s', self.name)
-        return ok
-
-    def reset(self):
-        logger.debug('Clear get callback %s', self.name)
-        self.gev.clear()
-
-    def __call__(self, exception=None):
-        logger.debug('Get callback in %s, exception=%s',
-                     self.name, exception)
-        if exception is None:
-            self.gev.set()
-
-
-def setup_pv(pvname, connect=True):
-    pv = pyca.capv(pvname)
-    pv.connect_cb = ConnectCallback(pvname)
-    pv.getevt_cb = GetCallback(pvname)
-    if connect:
-        pv.create_channel()
-        pv.connect_cb.wait(timeout=1)
-    return pv
